(feat) Modularize user data user for data gen
