+++ util/dataGen.js
@@ -2,6 +2,12 @@ module.export = (function() {
 
   'use strict';
 
+  const Nodal = require('Nodal');
+
+  const mockUserInfo = Nodal.require('mockData/mockUsers.js');
+
+  console.log(mockUserInfo);
+
   class User {
     constructor(username, email, password) {
       this.username = username;
@@ -10,51 +16,7 @@ module.export = (function() {
     }
   }
 
-  var usernames = [
-    'happy_sad',
-    'haveahappyday',
-    'SomethingNew',
-    '5mileys',
-    'takes2long',
-    'w8t4u',
-    'askme',
-    'Bidwell',
-    'massdebater',
-    'iluvmen',
-    'yomatey',
-    'idontknow',
-    'likme',
-    'TurtleCat',
-    'BlogWobbles',
-    'LuckyDusty',
-    'RumChicken',
-    'StonedTime',
-    'HackySacky',
-    'VillageIdiots'
-  ];
 
-  var names = [
-    'Emma',
-    'Noah',
-    'Olivia',
-    'Liam',
-    'Sophia',
-    'Mason',
-    'Isabella',
-    'Jacob',
-    'Ava',
-    'William',
-    'Mia',
-    'Ethan',
-    'Emily',
-    'Michael',
-    'Abigail',
-    'Alexander',
-    'Madison',
-    'James',
-    'Charlotte',
-    'Daniel'
-    ];
 
 /**
  * generate mock users
@@ -64,10 +26,10 @@ module.export = (function() {
  */
   var genUsers = (n, password) => {
     password = password || '11111';
-    n = n > usernames.length ? usernames.length : n;
-    let usernamesCopy = usernames.slice(0, n);
+    n = n > mockUserInfo.usernames.length ? mockUserInfo.usernames.length : n;
+    let usernames = mockUserInfo.usernames.slice(0, n);
     let users = usernamesCopy.map((username, i) => {
-      let name = names[i];
+      let name = mockUserInfo.names[i];
       let email = `${name.toLowerCase()}@gmail.com`;
       return new User(username, email, password);
     });
